make use of xrdesktop render-event
Previously we sent the frame done indicator to windows when the output
refreshed. This did not happen at the HMD refresh rate, but the monitor
refresh rate of the wlroots backedn window.

This commit causes windows to be updated at the native HMD refresh rate.
Additionally we lose the dependency on there being an output that
refreshes which will be useful for headless.